FE-1047: Use long instead of int for field Id

+++ MiX.Integrate.Shared/Entities/Assets/AdditionalDetailItem.cs
@@ -3,15 +3,15 @@
   /// <summary>Instance of a custom asset detail data point/summary>
   public class AdditionalDetailItem
   {
-    public AdditionalDetailItem(int id, string label, string value)
+    public AdditionalDetailItem(long id, string label, string value)
     {
       Id = id;
       Label = label;
       Value = value;
     }
 
-    /// <summary>Identifiedr of the custom detail</summary>
-    public int Id { get; set; }
+    /// <summary>Identifier of the custom detail</summary>
+    public long Id { get; set; }
 
     /// <summary>Name of the custom detail</summary>
     public string Label { get; set; }
